requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
yt-dlp
